refactor: extract users routes to routes/users.js
