+++ backend/server.js
@@ -3,7 +3,7 @@ require("dotenv").config();
 const express = require("express");
 const cors = require("cors");
 const rateLimit = require("express-rate-limit");
-const { db, auth } = require("./firebase");
+const { db } = require("./firebase");
 const admin = require("firebase-admin");
 const { removeUndefined, generateInviteCode, parseUTCToTimestamp, verifyAdmin, verifyFirebaseToken, checkCompanyAuthorization, resolveBooth, resolveApplicantResumePathOrUrl, requireCompanyResumeViewAccess, verifyRepOrOwner } = require("./helpers");
 const { GoogleGenerativeAI } = require("@google/generative-ai");
@@ -22,16 +22,6 @@ for (const envVar of requiredEnvVars) {
     process.exit(1);
   }
 }
-
-// --------------------------
-// STREAM CHAT SERVER CLIENT
-// --------------------------
-const { StreamChat } = require("stream-chat");
-
-const streamServer = StreamChat.getInstance(
-  process.env.STREAM_API_KEY,
-  process.env.STREAM_API_SECRET
-);
 
 const app = express();
 app.disable('x-powered-by'); // Disable Express version disclosure
@@ -120,6 +110,7 @@ app.use("/api", fairsRouter);
 app.use("/api", require("./routes/debug"));
 app.use("/api", require("./routes/stream"));
 app.use("/api", require("./routes/jobs"));
+app.use("/api", require("./routes/users"));
 
 /* ----------------------------------------------------
    UPLOAD RESUME TO FIREBASE STORAGE (via backend)
@@ -415,108 +406,6 @@ app.get("/api/get-booth-logo-url/:companyId", verifyFirebaseToken, async (req, r
   }
 });
 
-
-
-/* ----------------------------------------------------
-   REGISTER USER (Firestore + Stream upsert)
----------------------------------------------------- */
-app.post("/api/register-user", async (req, res) => {
-  try {
-    const { firstName, lastName, email, password, role, companyName } = req.body;
-
-    if (!email || !password || !role) {
-      return res
-        .status(400)
-        .send({ success: false, error: "Missing required fields" });
-    }
-
-    // Prevent administrator registration through public endpoint
-    if (role === "administrator") {
-      return res
-        .status(403)
-        .send({ success: false, error: "Administrator accounts cannot be created through public registration" });
-    }
-
-    // Create Firebase Auth user
-    const userRecord = await auth.createUser({
-      email,
-      password,
-      displayName: `${firstName || ""} ${lastName || ""}`.trim(),
-      emailVerified: true,
-    });
-
-    let companyId = null;
-    let plainInviteCode = null;
-
-    // Company owner setup
-    if (role === "companyOwner") {
-      plainInviteCode = generateInviteCode();
-
-      const companyRef = db.collection("companies").doc();
-      companyId = companyRef.id;
-
-      await companyRef.set(
-        removeUndefined({
-          companyId,
-          companyName,
-          ownerId: userRecord.uid,
-          inviteCode: plainInviteCode,
-          createdAt: admin.firestore.Timestamp.now(),
-        })
-      );
-    }
-
-    // Save user in Firestore (invite code is not stored on the user doc)
-    const docData = removeUndefined({
-      uid: userRecord.uid,
-      firstName,
-      lastName,
-      email,
-      role,
-      companyId,
-      companyName: role === "companyOwner" ? companyName : undefined,
-      emailVerified: true,
-      createdAt: admin.firestore.Timestamp.now(),
-    });
-
-    await db.collection("users").doc(userRecord.uid).set(docData);
-
-    // ---------------------------
-    // STREAM UPSERT USER
-    // ---------------------------
-    try {
-      const username = email.includes("@")
-        ? email.split("@")[0]
-        : email;
-
-      await streamServer.upsertUser({
-        id: userRecord.uid,
-        name: `${firstName || ""} ${lastName || ""}`.trim() || email,
-        email,
-        username,
-        firstName: firstName || "",
-        lastName: lastName || "",
-        role: "user", // Stream requires a valid role
-      });
-
-    } catch (error_) {
-      console.error("Stream upsert error:", error_);
-    }
-
-    res.send({
-      success: true,
-      user: {
-        uid: userRecord.uid,
-        email,
-        role,
-        companyId,
-      },
-    });
-  } catch (err) {
-    console.error("Error registering user:", err);
-    res.status(500).send({ success: false, error: err.message });
-  }
-});
 
 /* ----------------------------------------------------
    GET APPLICANT RESUME URL (for company reps/owners)
@@ -992,102 +881,6 @@ app.patch("/api/job-invitations/:id/status", async (req, res) => {
   } catch (err) {
     console.error("Error updating invitation status:", err);
     return res.status(500).json({ error: "Failed to update invitation status", details: err.message });
-  }
-});
-
-/* ----------------------------------------------------
-   GET LIST OF STUDENTS (for invitation UI)
----------------------------------------------------- */
-app.get("/api/students", async (req, res) => {
-  try {
-    const { userId, search, major, boothId } = req.query;
-
-    if (!userId) {
-      return res.status(400).json({ error: "User ID is required" });
-    }
-
-    // Verify user is representative, company owner, or admin
-    const authCheck = await verifyRepOrOwner(userId, null);
-    if (authCheck) {
-      return res.status(authCheck.status).json({ error: authCheck.error });
-    }
-
-    let students = [];
-
-    if (boothId) {
-      // Find students who visited this specific booth
-
-      // Get all students first
-      const allStudentsQuery = db.collection("users").where("role", "==", "student");
-      const studentsSnapshot = await allStudentsQuery.get();
-
-      // Check each student to see if they have a boothHistory document for this booth
-      const studentPromises = studentsSnapshot.docs.map(async (studentDoc) => {
-        const studentData = studentDoc.data();
-
-        // Check if this student has visited the booth
-        const boothHistoryDoc = await db.collection("users")
-          .doc(studentDoc.id)
-          .collection("boothHistory")
-          .doc(boothId)
-          .get();
-
-        if (boothHistoryDoc.exists) {
-          return {
-            id: studentDoc.id,
-            firstName: studentData.firstName || "",
-            lastName: studentData.lastName || "",
-            email: studentData.email || "",
-            major: studentData.major || "",
-          };
-        }
-        return null;
-      });
-
-      students = (await Promise.all(studentPromises)).filter(s => s !== null);
-      console.log(`Found ${students.length} students who visited booth`);
-    } else {
-      // Get all students (existing logic)
-      let query = db.collection("users").where("role", "==", "student");
-      const studentsSnapshot = await query.get();
-
-      students = studentsSnapshot.docs.map((doc) => {
-        const data = doc.data();
-        return {
-          id: doc.id,
-          firstName: data.firstName || "",
-          lastName: data.lastName || "",
-          email: data.email || "",
-          major: data.major || "",
-        };
-      });
-    }
-
-    // Apply search filter if provided
-    if (search?.trim()) {
-      const searchLower = search.toLowerCase().trim();
-      students = students.filter((student) => {
-        const fullName = `${student.firstName} ${student.lastName}`.toLowerCase();
-        return (
-          fullName.includes(searchLower) ||
-          student.email.toLowerCase().includes(searchLower)
-        );
-      });
-    }
-
-    // Apply major filter if provided
-    if (major?.trim()) {
-      const majorLower = major.toLowerCase().trim();
-      students = students.filter((student) =>
-        student.major.toLowerCase().includes(majorLower)
-      );
-    }
-
-    console.log(`Returning ${students.length} students`);
-    return res.json({ students });
-  } catch (err) {
-    console.error("Error fetching students:", err);
-    return res.status(500).json({ error: "Failed to fetch students", details: err.message });
   }
 });
 
@@ -1869,112 +1662,6 @@ app.post("/api/update-invite-code", async (req, res) => {
 });
 
 app.use("/api", require("./routes/companies"));
-
-/* ----------------------------------------------------
-   CREATE ADMIN ACCOUNT (Protected - requires secret key)
----------------------------------------------------- */
-app.post("/api/create-admin", async (req, res) => {
-  try {
-    const { firstName, lastName, email, password, adminSecret } = req.body;
-
-    if (!email || !password || !adminSecret) {
-      return res.status(400).json({ error: "Missing required fields" });
-    }
-
-    // Verify admin secret key
-    const expectedSecret = process.env.ADMIN_SECRET_KEY;
-    if (!expectedSecret) {
-      console.error("ADMIN_SECRET_KEY not set in environment variables");
-      return res.status(500).json({ error: "Server configuration error" });
-    }
-
-    if (adminSecret !== expectedSecret) {
-      return res.status(403).json({ error: "Invalid admin secret key" });
-    }
-
-    // Check if user already exists
-    let userRecord;
-    try {
-      userRecord = await auth.getUserByEmail(email);
-      // User exists, check if they're already an admin
-      const userDoc = await db.collection("users").doc(userRecord.uid).get();
-      if (userDoc.exists) {
-        const userData = userDoc.data();
-        if (userData.role === "administrator") {
-          return res.status(400).json({ error: "User is already an administrator" });
-        }
-        // Update existing user to admin
-        await db.collection("users").doc(userRecord.uid).update({
-          role: "administrator",
-        });
-        return res.json({
-          success: true,
-          message: "User upgraded to administrator",
-          uid: userRecord.uid
-        });
-      }
-    } catch (err) {
-      if (err.code !== "auth/user-not-found") {
-        throw err;
-      }
-      // User doesn't exist, create new one
-    }
-
-    // Create new Firebase Auth user
-    if (!userRecord) {
-      userRecord = await auth.createUser({
-        email,
-        password,
-        displayName: `${firstName || ""} ${lastName || ""}`.trim(),
-        emailVerified: true, // Auto-verify admin emails
-      });
-    }
-
-    // Save user in Firestore
-    const docData = removeUndefined({
-      uid: userRecord.uid,
-      firstName,
-      lastName,
-      email,
-      role: "administrator",
-      emailVerified: true,
-      createdAt: admin.firestore.Timestamp.now(),
-    });
-
-    await db.collection("users").doc(userRecord.uid).set(docData, { merge: true });
-
-    // Upsert to Stream Chat
-    try {
-      const username = email.includes("@")
-        ? email.split("@")[0]
-        : email;
-
-      await streamServer.upsertUser({
-        id: userRecord.uid,
-        name: `${firstName || ""} ${lastName || ""}`.trim() || email,
-        email,
-        username,
-        firstName: firstName || "",
-        lastName: lastName || "",
-        role: "user",
-      });
-    } catch (error_) {
-      console.error("STREAM UPSERT ERROR:", error_);
-    }
-
-    return res.json({
-      success: true,
-      user: {
-        uid: userRecord.uid,
-        email,
-        role: "administrator",
-      },
-    });
-  } catch (err) {
-    console.error("Error creating admin:", err);
-    return res.status(500).json({ error: err.message || "Failed to create admin account" });
-  }
-});
 
 app.post("/api/resume/tailor", verifyFirebaseToken, async (req, res) => {
   try {
